requirements.txt
streamlit
chromadb
sentence-transformers
numpy
langchain-groq
python-dotenv
pypdf
python-docx

